doc: Clarify handling of documents without headings
If a Markdown document has no headings defined in it,
`Inspect` will return a `TOC` without any items in it,
and `RenderList` will return nil.
Attempting to render this will cause a panic.

This is all already handled correctly by the
`Transformer` and the `Extender` (these are plug-and-play).

In manual mode, where users use `Inspect` and `RenderList` directly,
it makes sense to clarify that they need to handle the empty/nil cases.

Clarify this in the documentation where relevant.

Resolves #31