feat: Add MinDepth option (#40)
Adds a new 'MinDepth' option.
This option provides the ability to restrict the minimum header depth
to include in the toc.

Adds entries to the README and an option to the demo playground
for the new feature.

Testing:
Besides hand-written test cases, rapid-based property tests were added
to verify that the resulting TOC contains all headings.

Resolves #40